backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
alembic
httpx
python-jose[cryptography]
sendgrid
sqlalchemy-utils
python-multipart